==> 01-rbasicos.Rmd <==
# Fundamentos de R base

Este es un resumen de conceptos de R base útiles para los 
siguientes temas del curso.

Una revisión extensa de estos temas puede encontrarse en español
en: *R para principiantes* @bosco2018, o en inglés en: *An 
introduction to R* @douglas2022.

## Operaciones y variables

En R podemos realizar todo tipo de operaciones numéricas, 
ya sea en la consola o en un script. El resultado de la 
operación se desplegará automáticamente. 
```{r}
8934 + 452
2340 - 34
2344/45
356*23
# módulo: el residuo de la división
7394 %% 4
```

El caracter `#` nos permite agregar 
comentarios, es decir, código que no va a ejecutarse. 

```{r}
# Declarando una variable de tipo caracter
caracter <- "hola"
```

Si no sabemos los parámetros de una función o si necesitamos revisar qué es lo que hace, podemos hacer uso de la ayuda. En RStudio, esto se desplegará en el panel inferior derecho.

```{r}
?sd
help(sd)
```

### Variables

**Una variable es un par nombre, valor**. Creamos una variable y le asignamos un valor al declararla, mediante el operador de asignación 
`<-` e imprimimos su valor utilizando su nombre. Por ejemplo:

```{r}
mi_valor <- 42
mi_valor
```

Debemos elegir nombres cortos y descriptivos para nuestras variables. Esto ayudará a que nosotras y las personas que después lean nuestro código, puedan entenderlo fácilmente. 

Los nombres de variables pueden tener letras, números, puntos y guiones bajos, pero no pueden empezar con un número o con un punto seguido de un número. Además, debemos evitar usar palabras reservadas o nombres de funciones de R como  `NA`, `c`, `data`, etc.

En R se prefiere *snake case* (utilizar `_`) para nombres de variables con más de una palabra. 

Podemos realizar operaciones con las variables que declaramos
```{r}
mi_valor * 4.5
```


Y para actualizar el valor de nuestra variable, solamente tenemos que asignarlo nuevamente. No hay ninguna verificación sobre el tipo de valor con la que fue creada la variable inicialmente. 

```{r}
mi_valor <- "cuarenta y dos"
mi_valor
```

También podemos usar funciones de R base aplicadas a las variables que declaramos

```{r}
numero <- 89.63
floor(numero)
ceiling(numero)
```

Y almacenar el resultado en otra variable.

```{r}
fnum <- floor(numero)
```

## Errores comunes

Cuando tenemos un error en el nombre de la función que queremos utilizar o cuando no hemos cargado el paquete con la función correcta, observaremos el siguiente mensaje:

```{r error=TRUE}
flor(numero)
```

Cuando tenemos un error en la variable que utilizamos, en este caso como argumento de la función, el mensaje es el siguiente: 

```{r error=TRUE}
floor(numer)
```

## Tipos de datos

En R existen 6 tipos de datos básicos:

* Lógico (`logical`). Solamente tiene dos valores `TRUE` o `FALSE` y el tipo lógico especial `NA` para representar explícitamente valores faltantes o valores omitidos. 
* Entero (`integer`)
* Numérico (`numeric`)
* Caracteres (`character`)
* Números complejos (`complex`)
* Valores en bytes (`raw``)

No necesario especificar de forma explícita el tipo de datos de un objeto ya que R lo puede identificar automáticamente. Para conocer la clase a la que pertenece un objeto, 
se utiliza la función `class()`


```{r results='hide'}
num <- 88.66
class(num)
# [1] "numeric"
ent <- 80
class(ent)
# [1] "numeric"
ent <- 80L
class(ent)
# [1] "integer"
```

También podemos preguntar si un objeto es de alguna clase determinada

```{r results='hide'}
is.character(num)
# [1] FALSE
is.numeric(num)
# [1] TRUE"
```

Y convertir objetos de una clase a otra

```{r results='hide'}
as.numeric(TRUE)
# [1] "TRUE"
as.character(num)
# [1] "88.66"
```

Existen funciones que solamente pueden ser llamadas o ejecutadas con algún tipo específico de datos, ya que solo están definidas para ellos.

```{r results='hide',error=TRUE}
round(55.2)
# [1] 55
round("abc")
```

### Operadores lógicos 

Los operadores lógicos son muy importantes para la programación y tienen el mismo funcionamiento en todos los lenguajes de programación. Los  resultados de utilizar los principales operandos: `AND` (`&`), `OR` (`|`) y `NOT` (`!`) pueden ilustrarse 
en una tabla de verdad

```{r table2, echo=FALSE, message=FALSE, warnings=FALSE, results='asis'}
tabl <- "
|A      |B      |A&#124;B |A&B  | !A
|-------|-------|-------|-------|------
| FALSE | FALSE | FALSE | FALSE | TRUE
| FALSE | TRUE  | TRUE  | TRUE  | TRUE
| TRUE  | FALSE | TRUE  | FALSE | FALSE
| TRUE  | TRUE  | TRUE  | TRUE  | FALSE
"
cat(tabl) # output the table in a format good for HTML/PDF/docx conversion
```

### Cadenas de caracteres

Una función muy útil cuando trabajamos con cadenas de caracteres es `strsplit()`, que nos separa texto dado algún caracter. 

```{r}
cadena <- "uno;dos;tres;cuatro;cinco"
strsplit(cadena, split = ";")
```

Y para concatenar caracteres, números, variables, etc, utilizamos `paste()` o `paste0()`. La primera acepta un separador para concatenar y `paste0`, simplemente concatena todo. Además, el argumento para el caracter separador depende de los elementos del primer argumento. Si son objetos por separado usamos `sep` en `paste`

```{r}
paste("concatena", "palabras", "con", "espacios", sep = " ")
paste0("El valor guardado fue ", num, ". Sumando ", 5, " da ", num + 5)
```

Pero si los enviamos como un vector, usamos el argumento `collapse`

```{r}
paste(1:20, collapse = " ")
paste0(c("El valor guardado fue ", num, ". Sumando ", 5, " da ", num + 5), collapse = "")
```
