==> 04-ejemplos.Rmd <==
# Otros ejemplos de transformación

Esta sección contiene más ejemplos del uso de los verbos de dplyr. Cada resultado se guardará en un
nuevo archivo tsv para su uso posterior. 

En este caso utilizaremos el paquete `vroom` para leer y escribir datos.

```{r message=FALSE}
library(vroom)
library(dplyr)
library(janitor)
```

## Dataset de ANUIES 

Leemos el dataset de la Asociación Nacional de Universidades e Instituciones de Educación Superior y cambiamos las variables de `id` a caracteres. 

```{r message=FALSE}
anuies <- vroom("data/anuies.tsv")
anuies <- anuies %>%
  mutate(municipality_id = as.character(municipality_id), 
         state_id = as.character(state_id))

anuies <- anuies %>% 
  dplyr::mutate(state = replace(state, state == "Michoacán de Ocampo", "Michoacán"),
                state = replace(state, state == "Veracruz de Ignacio de la Llave", "Veracruz"),
                state = replace(state, state == "Coahuila de Zaragoza", "Coahuila"))
```

### Obtener el número de instituciones de educación superior en cada municipio

```{r message=FALSE}
anuies_by_mun <- anuies %>% 
  dplyr::select(municipality_id, municipality, state, institution) %>%
  ## El dataset también contiene áreas del conocimiento y otras variables, por lo que 
  ## nos aseguramos de contar una sola vez cada institución
  dplyr::distinct() %>%
  ## Agrupamos por variables geográficas y contamos
  dplyr::group_by(municipality_id, municipality, state) %>%
  ## Contamos las instituciones dado el agrupamiento y renombramos 
  ## la variable n
  dplyr::tally(name = "total")

head(anuies_by_mun)

vroom_write(anuies_by_mun, "data/anuies_by_mun.tsv")
```

### Obtener el número de instituciones de educación superior en cada municipio por grado académico

```{r message=FALSE}
anuies_by_mun_degree <- anuies %>% 
  dplyr::select(municipality_id, municipality, state, academic_degree, institution) %>%
  ## Contamos una sola vez cada institución por grado académico
  dplyr::distinct() %>%
  dplyr::group_by(municipality_id, municipality, state, academic_degree) %>%
  ## Contamos las instituciones dado el agrupamiento. 
  dplyr::tally(name = "total") 

head(anuies_by_mun_degree)
vroom_write(anuies_by_mun_degree, "data/anuies_by_mun_degree.tsv")
```

### Obtener el número de estudiantes de educación superior en cada municipio por grado académico y sexo {#ejanuies}

```{r}
estud_by_mun_sexo <- anuies %>% 
  dplyr::select(municipality_id, municipality, state, academic_degree, sex, students) %>%
  ## Hacemos un summary, para sumar el número de estudiantes
  ## agrupando por las variables geográficas, el grado académico y sexo
  dplyr::group_by(municipality_id, municipality, state, academic_degree, sex) %>%
  dplyr::summarise(estudiantes = sum(students)) 

head(estud_by_mun_sexo)
vroom_write(estud_by_mun_sexo, "data/estud_by_mun_sexo.tsv")
```

## Dataset de unidades de salud

```{r message=FALSE}
hospitales <- vroom::vroom("data/hospitales.tsv") %>%
  janitor::clean_names() %>%
  mutate(municipality_id = as.character(municipality_id), 
         state_id = as.character(state_id))

 hospitales <- hospitales %>% 
  dplyr::mutate(state = replace(state, state == "Michoacán de Ocampo", "Michoacán"),
                state = replace(state, state == "Veracruz de Ignacio de la Llave", "Veracruz"),
                state = replace(state, state == "Coahuila de Zaragoza", "Coahuila"))
```


### Obtener el número de unidades de salud en cada municipio

```{r message=FALSE}
hospitales_by_mun <- hospitales %>% 
  dplyr::select(municipality_id, municipality, state, clues) %>%
  ## Nos aseguramos de contar una sola vez cada institución
  dplyr::distinct() %>%
  ## Agrupamos por variables geográficas y contamos
  dplyr::group_by(municipality_id, municipality, state) %>%
  ## Contamos las instituciones dado el agrupamiento y renombramos 
  ## la variable n
  dplyr::tally(name = "total")

head(hospitales_by_mun)
vroom_write(hospitales_by_mun, "data/hospitales_by_mun.tsv")
```

### Obtener el número de unidades de salud en cada municipio por institución

```{r message=FALSE}
hospitales_by_mun_inst <- hospitales %>% 
  dplyr::select(municipality_id, municipality, state, clues, institution) %>%
  ## Nos aseguramos de contar una sola vez cada institución
  dplyr::distinct() %>%
  ## Agrupamos por variables geográficas y contamos
  dplyr::group_by(municipality_id, municipality, state, institution) %>%
  ## Contamos las instituciones dado el agrupamiento y renombramos 
  ## la variable n
  dplyr::tally(name = "total")

head(hospitales_by_mun_inst)
vroom_write(hospitales_by_mun_inst, "data/hospitales_by_mun_inst.tsv")
```

### Obtener el número de unidades de salud en cada municipio según el tipo de servicios que ofrecen

```{r message=FALSE}
hospitales_by_mun_serv <- hospitales %>% 
  dplyr::select(municipality_id, municipality, state, clues, establishment_type) %>%
  ## Nos aseguramos de contar una sola vez cada institución
  dplyr::distinct() %>%
  ## Agrupamos por variables geográficas y tipo de servicio y contamos
  dplyr::group_by(municipality_id, municipality, state, establishment_type) %>%
  ## Contamos las instituciones dado el agrupamiento y renombramos 
  ## la variable n
  dplyr::tally(name = "total")

head(hospitales_by_mun_serv)
vroom_write(hospitales_by_mun_serv, "data/hospitales_by_mun_serv.tsv")
```

### Obtener el número de camas en las unidades de salud en cada municipio por institución

```{r message=FALSE}
camas_by_mun_inst <- hospitales %>% 
  dplyr::select(municipality_id, municipality, state, institution, beds) %>%
  ## Agrupamos por variables geográficas y hacemos un summary
  dplyr::group_by(municipality_id, municipality, state, institution) %>%
  dplyr::summarise(camas = sum(beds)) %>%
  ## Eliminamos las entradas de las instituciones que no tienen camas
  ## en los municipios
  dplyr::filter(camas > 0)

head(camas_by_mun_inst)
vroom_write(camas_by_mun_inst, "data/camas_by_mun_inst.tsv")
```
